Bug in writing BitFieldArray things removed

+++ trunk/dstools-antlr/test/datascript/test/ParamBitFieldTest.java
@@ -20,6 +20,7 @@ import datascript.runtime.array.BitFieldArray;
 public class ParamBitFieldTest extends TestCase
 {
     private FileImageOutputStream os;
+    private String wFileName = "d:\\ParamBitFieldTest2.data";
     private String fileName = "parambitfieldtest.bin";
     private File file = new File(fileName);
 
@@ -44,14 +45,43 @@ public class ParamBitFieldTest extends TestCase
      */
     protected void tearDown() throws Exception
     {
-        file.delete();
+    	if (file.exists())
+    		file.delete();
+    }
+    
+    
+    private void checkBitField(VarBitField var, int size, 
+    		int numBits, int x, int y)
+    {
+        assertEquals(numBits, var.getNumBits());
+        
+        Coord c = var.getCoord();
+        assertEquals(x, c.getX().intValue());
+        assertEquals(y, c.getY().intValue());
+        
+        BitFieldArray array = c.getArray();
+        assertEquals(x, array.elementAt(0).intValue());
+        assertEquals(y, array.elementAt(1).intValue());
+        try
+        { 
+            int s = var.sizeof();
+            fail("RuntimeException expected");
+            assertEquals(size, s);
+        }
+        catch (RuntimeException exc)
+        {
+            // expected exception
+        }
     }
 
     /*
      * Test method for 'datascript.library.BitStreamReader.readByte()'
      */
-    private void writeBitField(int numBits, int x, int y) throws IOException
+    private int writeBitField(int numBits, int x, int y) throws IOException
     {
+    	if (file.exists())
+    		file.delete();
+
         os = new FileImageOutputStream(file);
         os.writeBits(numBits, 5);
         os.writeBits(x, numBits);
@@ -61,37 +91,42 @@ public class ParamBitFieldTest extends TestCase
         int size = (int) os.getStreamPosition();
         os.close();
 
-        VarBitField var = new VarBitField(fileName);
-        assertEquals(numBits, var.getNumBits());
-        Coord c = var.getCoord();
-        assertEquals(x, c.getX().intValue());
-        assertEquals(y, c.getY().intValue());
-        BitFieldArray array = c.getArray();
-        assertEquals(x, array.elementAt(0).intValue());
-        assertEquals(y, array.elementAt(1).intValue());
-        try
-        { 
-            int s = var.sizeof();
-            fail("RuntimeException expected");
-        }
-        catch (RuntimeException exc)
-        {
-            // expected exception
-        }
+        return size;
     }
         
     public void testBitField1() throws IOException
-    {
-        writeBitField(5, 10, 20);
+    {        
+        int size = writeBitField(5, 10, 20);
+        VarBitField var = new VarBitField(fileName);
+        checkBitField(var, size, 5, 10, 20);
+
+        var.write(wFileName);
+
+        VarBitField var2 = new VarBitField(wFileName);
+        checkBitField(var2, size, 5, 10, 20);
     }
 
     public void testBitField2() throws IOException
     {
-        writeBitField(5, 15, 27);
+    	int size = writeBitField(5, 15, 27);
+        VarBitField var = new VarBitField(fileName);
+        checkBitField(var, size, 5, 15, 27);
+        
+        var.write(wFileName);
+
+        VarBitField var2 = new VarBitField(wFileName);
+        checkBitField(var2, size, 5, 15, 27);
     }
 
     public void testBitField3() throws IOException
     {
-        writeBitField(10, 500, 1000);
+    	int size = writeBitField(10, 500, 1000);
+        VarBitField var = new VarBitField(fileName);
+        checkBitField(var, size, 10, 500, 1000);
+
+        var.write(wFileName);
+
+        VarBitField var2 = new VarBitField(wFileName);
+        checkBitField(var2, size, 10, 500, 1000);
     }
 }
